+++ src/containerapp/azext_containerapp/_validators.py
@@ -8,6 +8,7 @@ from azure.cli.core.azclierror import (ValidationError, ResourceNotFoundError, I
                                        MutuallyExclusiveArgumentError)
 from msrestazure.tools import is_valid_resource_id
 from knack.log import get_logger
+import re
 
 from ._clients import ContainerAppClient
 from ._ssh_utils import ping_container_app
@@ -36,6 +37,14 @@ def _is_number(s):
         return True
     except ValueError:
         return False
+
+
+def validate_revision_suffix(value):
+    if value is not None:
+        matched = re.match(r"^[a-z](?!.*-{2})([-a-z0-9]*[a-z0-9])?$", value)
+        if not matched:
+            raise ValidationError(f"Invalid Container App revision name '{value}'. A revision name must consist of lower case alphanumeric characters or '-', start with a letter, end with an alphanumeric character and cannot have '--'.")
+
 
 
 def validate_memory(namespace):